Incorporate quadratic prefactor into coefficients
Instead of scaling the quadratic coefficients each time they are used as
part of the Hamiltonian, they are divided by 2 when reading from a
system configuration JSON file and multiplied by 2 when writing to one.
Thus, from the perspective of the JSON files, nothing has changed.
Internally, however, the implicit Hamiltonian no longer has a quadratic
prefactor of 1/2.